RESTAPI/requirements/requirements.txt
-r test-requirements.txt
appdirs==1.4.0
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==1.5.0
certifi==2017.1.23
configparser==3.5.0
decorator==4.0.11
Django==1.10.5
django-debug-toolbar==1.2.2
django-extensions==1.7.6
djangorestframework==3.5.4
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
gnureadline==6.3.3
html5lib==0.9999999
ipdb==0.10.2
ipykernel==4.5.2
ipython==5.2.2
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==0.23
mistune==0.7.3
nbconvert==5.1.1
nbformat==4.3.0
notebook==4.4.1
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
plotly==2.0.2
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.13.0
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.2
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.2
wcwidth==0.1.7
widgetsnbextension==1.2.6
XlsxWriter==0.9.6
